requirements.txt
altair==4.2.2
matplotlib==3.5.2
numpy==1.24.4
openai==0.26.5
pandas==2.0.2
Pillow==10.3.0
plotly==5.7.0
PyPDF2==3.0.1
requests
scikit_learn==1.2.0
scipy
seaborn==0.13.2
streamlit
pyvis
torch
g4f
curl-cffi
streamlit-wallet-connect